requirements.txt
pandas
numpy
torch
pillow
